tests: move signtxVerify fixtures into separate file
